wip: refactor to better file organization
