+++ crates/duckdb/Cargo.toml
@@ -22,6 +22,8 @@ default = []
 bundled = ["libduckdb-sys/bundled"]
 json = ["libduckdb-sys/json", "bundled"]
 parquet = ["libduckdb-sys/parquet", "bundled"]
+vscalar = []
+vscalar-arrow = []
 vtab = []
 vtab-loadable = ["vtab", "duckdb-loadable-macros"]
 vtab-excel = ["vtab", "calamine"]
