+++ crates/duckdb/src/vtab/mod.rs
@@ -2,8 +2,6 @@ use crate::{error::Error, inner_connection::InnerConnection, Connection, Result}
 
 use super::{ffi, ffi::duckdb_free};
 use std::ffi::c_void;
-use std::fmt::Debug;
-use std::sync::Arc;
 
 mod function;
 mod value;
@@ -19,12 +17,7 @@ pub use self::arrow::{
 #[cfg(feature = "vtab-excel")]
 mod excel;
 
-use ::arrow::array::{Array, RecordBatch};
-use ::arrow::datatypes::DataType;
-use arrow::{data_chunk_to_arrow, write_arrow_array_to_vector, WritableVector};
 pub use function::{BindInfo, InitInfo, TableFunction, TableFunctionInfo};
-use function::{ScalarFunction, ScalarFunctionInfo, ScalarFunctionSet};
-use libduckdb_sys::duckdb_vector;
 pub use value::Value;
 
 use crate::core::{DataChunkHandle, LogicalTypeHandle, LogicalTypeId};
@@ -163,179 +156,6 @@ where
     }
 }
 
-/// Duckdb scalar function parameters
-pub enum Parameters {
-    /// Exact parameters
-    Exact(Vec<LogicalTypeHandle>),
-    /// Variadic parameters
-    Variadic(LogicalTypeHandle),
-}
-
-/// Duckdb scalar function signature
-pub struct ScalarFunctionSignature {
-    parameters: Option<Parameters>,
-    return_type: LogicalTypeHandle,
-}
-
-impl ScalarFunctionSignature {
-    /// Create an exact function signature
-    pub fn exact(params: Vec<LogicalTypeHandle>, return_type: LogicalTypeHandle) -> Self {
-        ScalarFunctionSignature {
-            parameters: Some(Parameters::Exact(params)),
-            return_type,
-        }
-    }
-
-    /// Create a variadic function signature
-    pub fn variadic(param: LogicalTypeHandle, return_type: LogicalTypeHandle) -> Self {
-        ScalarFunctionSignature {
-            parameters: Some(Parameters::Variadic(param)),
-            return_type,
-        }
-    }
-}
-
-impl ScalarFunctionSignature {
-    fn register_with_scalar(&self, f: &ScalarFunction) {
-        f.set_return_type(&self.return_type);
-
-        match &self.parameters {
-            Some(Parameters::Exact(params)) => {
-                for param in params.iter() {
-                    f.add_parameter(param);
-                }
-            }
-            Some(Parameters::Variadic(param)) => {
-                f.add_variadic_parameter(param);
-            }
-            None => {
-                // do nothing
-            }
-        }
-    }
-}
-
-/// Duckdb scalar function trait
-pub trait VScalar: Sized {
-    /// State that persists across invocations of the scalar function (the lifetime of the connection)
-    type State: Default;
-    /// The actual function
-    ///
-    /// # Safety
-    ///
-    /// This function is unsafe because it:
-    ///
-    /// - Dereferences multiple raw pointers (`func``).
-    ///
-    unsafe fn invoke(
-        state: &Self::State,
-        input: &mut DataChunkHandle,
-        output: &mut dyn WritableVector,
-    ) -> Result<(), Box<dyn std::error::Error>>;
-
-    /// The possible signatures of the scalar function
-    fn signatures() -> Vec<ScalarFunctionSignature>;
-}
-
-pub enum ArrowParams {
-    Exact(Vec<DataType>),
-    Variadic(DataType),
-}
-
-impl AsRef<[DataType]> for ArrowParams {
-    fn as_ref(&self) -> &[DataType] {
-        match self {
-            ArrowParams::Exact(params) => params.as_ref(),
-            ArrowParams::Variadic(param) => std::slice::from_ref(param),
-        }
-    }
-}
-
-impl From<ArrowParams> for Parameters {
-    fn from(params: ArrowParams) -> Self {
-        match params {
-            ArrowParams::Exact(params) => Parameters::Exact(
-                params
-                    .into_iter()
-                    .map(|v| LogicalTypeId::try_from(&v).expect("type should be converted").into())
-                    .collect(),
-            ),
-            ArrowParams::Variadic(param) => Parameters::Variadic(
-                LogicalTypeId::try_from(&param)
-                    .expect("type should be converted")
-                    .into(),
-            ),
-        }
-    }
-}
-
-pub struct ArrowFunctionSignature {
-    pub parameters: Option<ArrowParams>,
-    pub return_type: DataType,
-}
-
-impl ArrowFunctionSignature {
-    pub fn exact(params: Vec<DataType>, return_type: DataType) -> Self {
-        ArrowFunctionSignature {
-            parameters: Some(ArrowParams::Exact(params)),
-            return_type,
-        }
-    }
-}
-
-/// blah
-pub trait ArrowScalar: Sized {
-    /// blah
-    type State: Default;
-
-    /// blah
-    fn invoke(info: &Self::State, input: RecordBatch) -> Result<Arc<dyn Array>, Box<dyn std::error::Error>>;
-
-    /// blah
-    fn signatures() -> Vec<ArrowFunctionSignature>;
-}
-
-impl<T> VScalar for T
-where
-    T: ArrowScalar,
-    T::State: Debug,
-{
-    type State = T::State;
-
-    unsafe fn invoke(
-        info: &Self::State,
-        input: &mut DataChunkHandle,
-        out: &mut dyn WritableVector,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        let array = T::invoke(info, data_chunk_to_arrow(input)?)?;
-        write_arrow_array_to_vector(&array, out)
-    }
-
-    fn signatures() -> Vec<ScalarFunctionSignature> {
-        T::signatures()
-            .into_iter()
-            .map(|sig| ScalarFunctionSignature {
-                parameters: sig.parameters.map(Into::into),
-                return_type: LogicalTypeId::try_from(&sig.return_type)
-                    .expect("type should be converted")
-                    .into(),
-            })
-            .collect()
-    }
-}
-
-unsafe extern "C" fn scalar_func<T>(info: duckdb_function_info, input: duckdb_data_chunk, mut output: duckdb_vector)
-where
-    T: VScalar,
-{
-    let info = ScalarFunctionInfo::from(info);
-    let mut input = DataChunkHandle::new_unowned(input);
-    let result = T::invoke(info.get_scalar_extra_info(), &mut input, &mut output);
-    if let Err(e) = result {
-        info.set_error(&e.to_string());
-    }
-}
-
 impl Connection {
     /// Register the given TableFunction with the current db
     #[inline]
@@ -355,23 +175,6 @@ impl Connection {
         }
         self.db.borrow_mut().register_table_function(table_function)
     }
-
-    /// Register the given ScalarFunction with the current db
-    #[inline]
-    pub fn register_scalar_function<S: VScalar>(&self, name: &str) -> Result<()>
-    where
-        S::State: Debug,
-    {
-        let set = ScalarFunctionSet::new(name);
-        for signature in S::signatures() {
-            let scalar_function = ScalarFunction::new(name)?;
-            signature.register_with_scalar(&scalar_function);
-            scalar_function.set_function(Some(scalar_func::<S>));
-            scalar_function.set_extra_info::<S::State>();
-            set.add_function(scalar_function)?;
-        }
-        self.db.borrow_mut().register_scalar_function_set(set)
-    }
 }
 
 impl InnerConnection {
@@ -384,11 +187,6 @@ impl InnerConnection {
             }
         }
         Ok(())
-    }
-
-    /// Register the given ScalarFunction with the current db
-    pub fn register_scalar_function_set(&mut self, f: ScalarFunctionSet) -> Result<()> {
-        f.register_with_connection(self.con)
     }
 }
 
@@ -506,239 +304,6 @@ mod test {
         }
     }
 
-    struct HelloScalarArrow {}
-
-    impl ArrowScalar for HelloScalarArrow {
-        type State = ();
-
-        fn invoke(_: &Self::State, input: RecordBatch) -> Result<Arc<dyn Array>, Box<dyn std::error::Error>> {
-            let name = input.column(0).as_any().downcast_ref::<StringArray>().unwrap();
-            let result = name.iter().map(|v| format!("Hello {}", v.unwrap())).collect::<Vec<_>>();
-            Ok(Arc::new(StringArray::from(result)))
-        }
-
-        fn signatures() -> Vec<ArrowFunctionSignature> {
-            vec![ArrowFunctionSignature::exact(vec![DataType::Utf8], DataType::Utf8)]
-        }
-    }
-
-    #[derive(Debug)]
-    struct MockState {
-        info: String,
-    }
-
-    impl Default for MockState {
-        fn default() -> Self {
-            MockState {
-                info: "some meta".to_string(),
-            }
-        }
-    }
-
-    impl Drop for MockState {
-        fn drop(&mut self) {
-            println!("dropped meta");
-        }
-    }
-
-    struct ArrowMultiplyScalar {}
-
-    impl ArrowScalar for ArrowMultiplyScalar {
-        type State = MockState;
-
-        fn invoke(_: &Self::State, input: RecordBatch) -> Result<Arc<dyn Array>, Box<dyn std::error::Error>> {
-            let a = input
-                .column(0)
-                .as_any()
-                .downcast_ref::<::arrow::array::Float32Array>()
-                .unwrap();
-
-            let b = input
-                .column(1)
-                .as_any()
-                .downcast_ref::<::arrow::array::Float32Array>()
-                .unwrap();
-
-            let result = a
-                .iter()
-                .zip(b.iter())
-                .map(|(a, b)| a.unwrap() * b.unwrap())
-                .collect::<Vec<_>>();
-            Ok(Arc::new(::arrow::array::Float32Array::from(result)))
-        }
-
-        fn signatures() -> Vec<ArrowFunctionSignature> {
-            vec![ArrowFunctionSignature::exact(
-                vec![DataType::Float32, DataType::Float32],
-                DataType::Float32,
-            )]
-        }
-    }
-
-    // accepts a string or a number and parses to int and multiplies by 2
-    struct ArrowOverloaded {}
-
-    impl ArrowScalar for ArrowOverloaded {
-        type State = MockState;
-
-        fn invoke(s: &Self::State, input: RecordBatch) -> Result<Arc<dyn Array>, Box<dyn std::error::Error>> {
-            assert_eq!("some meta", s.info);
-
-            let a = input.column(0);
-            let b = input.column(1);
-
-            let result = match a.data_type() {
-                DataType::Utf8 => {
-                    let a = a
-                        .as_any()
-                        .downcast_ref::<::arrow::array::StringArray>()
-                        .unwrap()
-                        .iter()
-                        .map(|v| v.unwrap().parse::<f32>().unwrap())
-                        .collect::<Vec<_>>();
-                    let b = b
-                        .as_any()
-                        .downcast_ref::<::arrow::array::Float32Array>()
-                        .unwrap()
-                        .iter()
-                        .map(|v| v.unwrap())
-                        .collect::<Vec<_>>();
-                    a.iter().zip(b.iter()).map(|(a, b)| a * b).collect::<Vec<_>>()
-                }
-                DataType::Float32 => {
-                    let a = a
-                        .as_any()
-                        .downcast_ref::<::arrow::array::Float32Array>()
-                        .unwrap()
-                        .iter()
-                        .map(|v| v.unwrap())
-                        .collect::<Vec<_>>();
-                    let b = b
-                        .as_any()
-                        .downcast_ref::<::arrow::array::Float32Array>()
-                        .unwrap()
-                        .iter()
-                        .map(|v| v.unwrap())
-                        .collect::<Vec<_>>();
-                    a.iter().zip(b.iter()).map(|(a, b)| a * b).collect::<Vec<_>>()
-                }
-                _ => panic!("unsupported type"),
-            };
-
-            Ok(Arc::new(::arrow::array::Float32Array::from(result)))
-        }
-
-        fn signatures() -> Vec<ArrowFunctionSignature> {
-            vec![
-                ArrowFunctionSignature::exact(vec![DataType::Utf8, DataType::Float32], DataType::Float32),
-                ArrowFunctionSignature::exact(vec![DataType::Float32, DataType::Float32], DataType::Float32),
-            ]
-        }
-    }
-
-    struct ErrorScalar {}
-
-    impl VScalar for ErrorScalar {
-        type State = ();
-
-        unsafe fn invoke(
-            _: &Self::State,
-            input: &mut DataChunkHandle,
-            _: &mut dyn WritableVector,
-        ) -> Result<(), Box<dyn std::error::Error>> {
-            let mut msg = input.flat_vector(0).as_slice_with_len::<duckdb_string_t>(input.len())[0];
-            let string = DuckString::new(&mut msg).as_str();
-            Err(format!("Error: {}", string).into())
-        }
-
-        fn signatures() -> Vec<ScalarFunctionSignature> {
-            vec![ScalarFunctionSignature::exact(
-                vec![LogicalTypeId::Varchar.into()],
-                LogicalTypeId::Varchar.into(),
-            )]
-        }
-    }
-
-    #[derive(Debug)]
-    struct TestState {
-        #[allow(dead_code)]
-        inner: i32,
-    }
-
-    impl Default for TestState {
-        fn default() -> Self {
-            TestState { inner: 42 }
-        }
-    }
-
-    struct EchoScalar {}
-
-    impl VScalar for EchoScalar {
-        type State = TestState;
-
-        unsafe fn invoke(
-            s: &Self::State,
-            input: &mut DataChunkHandle,
-            output: &mut dyn WritableVector,
-        ) -> Result<(), Box<dyn std::error::Error>> {
-            assert_eq!(s.inner, 42);
-            let values = input.flat_vector(0);
-            let values = values.as_slice_with_len::<duckdb_string_t>(input.len());
-            let strings = values
-                .iter()
-                .map(|ptr| DuckString::new(&mut { *ptr }).as_str().to_string())
-                .take(input.len());
-            let output = output.flat_vector();
-            for s in strings {
-                output.insert(0, s.to_string().as_str());
-            }
-            Ok(())
-        }
-
-        fn signatures() -> Vec<ScalarFunctionSignature> {
-            vec![ScalarFunctionSignature::exact(
-                vec![LogicalTypeId::Varchar.into()],
-                LogicalTypeId::Varchar.into(),
-            )]
-        }
-    }
-
-    struct Repeat {}
-
-    impl VScalar for Repeat {
-        type State = ();
-
-        unsafe fn invoke(
-            _: &Self::State,
-            input: &mut DataChunkHandle,
-            output: &mut dyn WritableVector,
-        ) -> Result<(), Box<dyn std::error::Error>> {
-            let output = output.flat_vector();
-            let counts = input.flat_vector(1);
-            let values = input.flat_vector(0);
-            let values = values.as_slice_with_len::<duckdb_string_t>(input.len());
-            let strings = values
-                .iter()
-                .map(|ptr| DuckString::new(&mut { *ptr }).as_str().to_string());
-            let counts = counts.as_slice_with_len::<i32>(input.len());
-            for (count, value) in counts.iter().zip(strings).take(input.len()) {
-                output.insert(0, value.repeat((*count) as usize).as_str());
-            }
-
-            Ok(())
-        }
-
-        fn signatures() -> Vec<ScalarFunctionSignature> {
-            vec![ScalarFunctionSignature::exact(
-                vec![
-                    LogicalTypeHandle::from(LogicalTypeId::Varchar),
-                    LogicalTypeHandle::from(LogicalTypeId::Integer),
-                ],
-                LogicalTypeHandle::from(LogicalTypeId::Varchar),
-            )]
-        }
-    }
-
     #[test]
     fn test_table_function() -> Result<(), Box<dyn Error>> {
         let conn = Connection::open_in_memory()?;
@@ -763,137 +328,8 @@ mod test {
         Ok(())
     }
 
-    #[test]
-    fn test_scalar() -> Result<(), Box<dyn Error>> {
-        let conn = Connection::open_in_memory()?;
-        conn.register_scalar_function::<EchoScalar>("echo")?;
-
-        let mut stmt = conn.prepare("select echo('hi') as hello")?;
-        let mut rows = stmt.query([])?;
-
-        while let Some(row) = rows.next()? {
-            let hello: String = row.get(0)?;
-            assert_eq!(hello, "hi");
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_scalar_error() -> Result<(), Box<dyn Error>> {
-        let conn = Connection::open_in_memory()?;
-        conn.register_scalar_function::<ErrorScalar>("error_udf")?;
-
-        let mut stmt = conn.prepare("select error_udf('blurg') as hello")?;
-        if let Err(err) = stmt.query([]) {
-            assert!(err.to_string().contains("Error: blurg"));
-        } else {
-            panic!("Expected an error");
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_arrow_scalar() -> Result<(), Box<dyn Error>> {
-        let conn = Connection::open_in_memory()?;
-        conn.register_scalar_function::<HelloScalarArrow>("hello")?;
-
-        let batches = conn
-            .prepare("select hello('foo') as hello from range(10)")?
-            .query_arrow([])?
-            .collect::<Vec<_>>();
-
-        for batch in batches.iter() {
-            let array = batch.column(0);
-            let array = array.as_any().downcast_ref::<::arrow::array::StringArray>().unwrap();
-            for i in 0..array.len() {
-                assert_eq!(array.value(i), format!("Hello foo"));
-            }
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_arrow_scalar_multiply() -> Result<(), Box<dyn Error>> {
-        let conn = Connection::open_in_memory()?;
-        conn.register_scalar_function::<ArrowMultiplyScalar>("multiply_udf")?;
-
-        let batches = conn
-            .prepare("select multiply_udf(3.0, 2.0) as mult_result from range(10)")?
-            .query_arrow([])?
-            .collect::<Vec<_>>();
-
-        for batch in batches.iter() {
-            let array = batch.column(0);
-            let array = array.as_any().downcast_ref::<::arrow::array::Float32Array>().unwrap();
-            for i in 0..array.len() {
-                assert_eq!(array.value(i), 6.0);
-            }
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_repeat_scalar() -> Result<(), Box<dyn Error>> {
-        let conn = Connection::open_in_memory()?;
-        conn.register_scalar_function::<Repeat>("nobie_repeat")?;
-
-        let batches = conn
-            .prepare("select nobie_repeat('Ho ho ho 🎅🎄', 3) as message from range(5)")?
-            .query_arrow([])?
-            .collect::<Vec<_>>();
-
-        for batch in batches.iter() {
-            let array = batch.column(0);
-            let array = array.as_any().downcast_ref::<::arrow::array::StringArray>().unwrap();
-            for i in 0..array.len() {
-                assert_eq!(array.value(i), "Ho ho ho 🎅🎄Ho ho ho 🎅🎄Ho ho ho 🎅🎄");
-            }
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_multiple_signatures_scalar() -> Result<(), Box<dyn Error>> {
-        let conn = Connection::open_in_memory()?;
-        conn.register_scalar_function::<ArrowOverloaded>("multi_sig_udf")?;
-
-        let batches = conn
-            .prepare("select multi_sig_udf('3', 5) as message from range(2)")?
-            .query_arrow([])?
-            .collect::<Vec<_>>();
-
-        for batch in batches.iter() {
-            let array = batch.column(0);
-            let array = array.as_any().downcast_ref::<::arrow::array::Float32Array>().unwrap();
-            for i in 0..array.len() {
-                assert_eq!(array.value(i), 15.0);
-            }
-        }
-
-        let batches = conn
-            .prepare("select multi_sig_udf(12, 10) as message from range(2)")?
-            .query_arrow([])?
-            .collect::<Vec<_>>();
-
-        for batch in batches.iter() {
-            let array = batch.column(0);
-            let array = array.as_any().downcast_ref::<::arrow::array::Float32Array>().unwrap();
-            for i in 0..array.len() {
-                assert_eq!(array.value(i), 120.0);
-            }
-        }
-
-        Ok(())
-    }
-
-    use ::arrow::array::StringArray;
     #[cfg(feature = "vtab-loadable")]
     use duckdb_loadable_macros::duckdb_entrypoint;
-    use libduckdb_sys::duckdb_string_t;
 
     // this function is never called, but is still type checked
     // Exposes a extern C function named "libhello_ext_init" in the compiled dynamic library,
